+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Change log
 For detailed information check [Blogophon's releases on Github](https://github.com/fboes/blogophon/releases).
 
 * 💊 Fix ARIA handling of gallery links
+* 💊 Fix internal download links for teasers
 * 🎁 Add Gopher server support
 * 🎁 Add Spotify embedding
 * 🎁 JSON-LD added